+++ src/php/Command/CommandConfig.php
@@ -14,11 +14,13 @@ final class CommandConfig extends AbstractConfig
     public const TEST_DIRS = 'test-dirs';
     public const VENDOR_DIR = 'vendor-dir';
     public const OUTPUT = 'out';
+    public const ERROR_LOG_FILE = 'error-log-file';
 
     private const DEFAULT_VENDOR_DIR = 'vendor';
     private const DEFAULT_SRC_DIRS = ['src'];
     private const DEFAULT_TEST_DIRS = ['tests'];
     private const DEFAULT_OUTPUT_DIR = 'out';
+    private const DEFAULT_ERROR_LOG_FILE = 'data/error.log';
 
     public function getCodeDirs(): CodeDirectories
     {
@@ -34,5 +36,10 @@ final class CommandConfig extends AbstractConfig
     public function getVendorDir(): string
     {
         return (string)$this->get(self::VENDOR_DIR, self::DEFAULT_VENDOR_DIR);
+    }
+
+    public function getErrorLogFile(): string
+    {
+        return (string)$this->get(self::ERROR_LOG_FILE, self::DEFAULT_ERROR_LOG_FILE);
     }
 }