+++ src/php/Command/CommandFacade.php
@@ -5,9 +5,9 @@ declare(strict_types=1);
 namespace Phel\Command;
 
 use Gacela\Framework\AbstractFacade;
+use Phel\Command\Domain\Handler\ExceptionHandler;
 use Phel\Command\Domain\Shared\Exceptions\ExceptionPrinterInterface;
 use Phel\Compiler\Domain\Exceptions\AbstractLocatedException;
-use Phel\Compiler\Domain\Exceptions\CompilerException;
 use Phel\Compiler\Domain\Parser\ReadModel\CodeSnippet;
 use Symfony\Component\Console\Output\OutputInterface;
 use Throwable;
@@ -50,15 +50,7 @@ final class CommandFacade extends AbstractFacade implements CommandFacadeInterfa
 
     public function registerExceptionHandler(): void
     {
-        $exceptionPrinter = $this->getExceptionPrinter();
-
-        set_exception_handler(static function (Throwable $exception) use ($exceptionPrinter): void {
-            if ($exception instanceof CompilerException) {
-                $exceptionPrinter->printException($exception->getNestedException(), $exception->getCodeSnippet());
-            } else {
-                $exceptionPrinter->printStackTrace($exception);
-            }
-        });
+        $this->createExceptionHandler()->register();
     }
 
     /**
@@ -116,5 +108,12 @@ final class CommandFacade extends AbstractFacade implements CommandFacadeInterfa
         return $this->getFactory()
             ->getPhpConfigReader()
             ->read($absolutePath);
+    }
+
+    private function createExceptionHandler(): ExceptionHandler
+    {
+        return new ExceptionHandler(
+            $this->getExceptionPrinter(),
+        );
     }
 }