+++ src/php/Command/Domain/Shared/Exceptions/ExceptionPrinterInterface.php
@@ -10,6 +10,8 @@ use Throwable;
 
 interface ExceptionPrinterInterface
 {
+    public function printError(string $error): void;
+
     public function printException(AbstractLocatedException $e, CodeSnippet $codeSnippet): void;
 
     public function getExceptionString(AbstractLocatedException $e, CodeSnippet $codeSnippet): string;